move isConnectedStateFlow to the base module
