+++ ui/base/src/main/AndroidManifest.xml
@@ -1,6 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
-
     <application>
         <activity android:name="org.cru.godtools.base.ui.firebase.DynamicLinksSpringboardActivity"
             android:exported="true">
